Small refactors around making the ErrorPanel more generic
